SnnsRObject$createNet {RSNNS}	R Documentation

Create a layered network

Description

This function creates a layered network in the given SnnsR object. This is an SnnsR low-level function. You may want to have a look at SnnsR-class to find out how to properly use it.

Usage

## S4 method for signature 'SnnsR'
createNet(unitsPerLayer, fullyConnectedFeedForward = TRUE, iNames = NULL, oNames = NULL)

Arguments

unitsPerLayer
a vector of integers that represents the number of units in each layer, including input and output layer

fullyConnectedFeedForward
if TRUE, the network is fully connected as a feed-forward network. If FALSE, no connections are created

iNames
names of input units

oNames
names of output units

See Also

SnnsR-class

Examples

obj1 <- SnnsRObjectFactory()
obj1$createNet(c(2,2), FALSE)
obj1$getUnitDefinitions()

obj2 <- SnnsRObjectFactory()
obj2$createNet(c(8,5,5,2), TRUE)
obj2$getUnitDefinitions()

[Package RSNNS version 0.4-7 Index]