SnnsRObject$createPatSet {RSNNS}	R Documentation

Create a pattern set

Description

SnnsR low-level function to create a pattern set in the SNNS kernel from the values given, so that they are available in the SNNS kernel for use.

Usage

## S4 method for signature 'SnnsR'
createPatSet(inputs, targets)

Arguments

inputs
the input values

targets
the target values

Value

a list with elements err and set_no. The latter one identifies the pattern set within the SnnsR-class object

[Package RSNNS version 0.4-7 Index]