SnnsRObject$getAllUnitsTType {RSNNS}	R Documentation

Get all units in the net of a certain ttype.

Description

SnnsR low-level function to get all units in the net of a certain ttype. Possible ttype defined by SNNS are, among others: "UNIT_OUTPUT", "UNIT_INPUT", and "UNIT_HIDDEN". For a full list, call RSNNS:::SnnsDefines$topologicalUnitTypes As this is an SnnsR low-level function, you may want to have a look at SnnsR-class to find out how to properly use it.

Usage

## S4 method for signature 'SnnsR'
getAllUnitsTType(ttype)

Arguments

ttype
a string containing the ttype.

Value

a vector with integer numbers identifying the units.

See Also

SnnsRObject$getAllOutputUnits, SnnsRObject$getAllInputUnits, SnnsRObject$getAllHiddenUnits

[Package RSNNS version 0.4-7 Index]