SnnsRObject$getWeightMatrix {RSNNS}	R Documentation

Get the weight matrix between two sets of units

Description

SnnsR low-level function to get the weight matrix between two sets of units.

Usage

## S4 method for signature 'SnnsR'
getWeightMatrix(unitsSource, unitsTarget, setDimNames)

Arguments

unitsSource
a vector with numbers identifying the source units

unitsTarget
a vector with numbers identifying the target units

setDimNames
indicates, whether names of units are extracted and set as row/col names in the weight matrix

Value

the weight matrix between the two sets of neurons

See Also

SnnsRObject$getAllUnitsTType

[Package RSNNS version 0.4-7 Index]