SnnsRObject$initializeNet {RSNNS}	R Documentation

Initialize the network

Description

This SnnsR low-level function masks the SNNS kernel function of the same name to allow for both giving the initialization function directly in the call or to use the one that is currently set.

Usage

## S4 method for signature 'SnnsR'
initializeNet(parameterInArray, initFunc)

Arguments

parameterInArray
the parameters of the initialization function

initFunc
the name of the initialization function

[Package RSNNS version 0.4-7 Index]