SnnsRObject$predictCurrPatSet {RSNNS}	R Documentation

Predict values with a trained net

Description

SnnsR low-level function to predict values with a trained net.

Usage

## S4 method for signature 'SnnsR'
predictCurrPatSet(outputMethod="reg_class", updateFuncParams=c(0.0))

Arguments

outputMethod
is passed to SnnsRObject$whereAreResults

updateFuncParams
parameters passed to the networks update function

Details

This function has to be used embedded in a step of loading and afterwards removing the patterns into the SnnsR-class object. As SNNS only supports 2 pattern sets in parallel, removing unneeded pattern sets is quite important.

Value

the predicted values

[Package RSNNS version 0.4-7 Index]