SnnsRObject$setTTypeUnitsActFunc {RSNNS}	R Documentation

Set the activation function for all units of a certain ttype.

Description

The function uses the function SnnsRObject$getAllUnitsTType to find all units of a certain ttype, and sets the activation function of all these units to the given activation function.

Usage

## S4 method for signature 'SnnsR'
setTTypeUnitsActFunc(ttype, act_func)

Arguments

ttype
a string containing the ttype.

act_func
the name of the activation function to set.

See Also

SnnsRObject$getAllUnitsTType

Examples

## Not run: SnnsRObject$setTTypeUnitsActFunc("UNIT_HIDDEN", "Act_Logistic")

[Package RSNNS version 0.4-7 Index]