SnnsRObject$setUnitDefaults {RSNNS}	R Documentation

Set the unit defaults

Description

This SnnsR low-level function masks the SNNS kernel function of the same name to allow both for giving the parameters directly or as a vector. If the second parameter, bias, is missing, it is assumed that the first parameter should be interpreted as a vector containing all parameters.

Usage

## S4 method for signature 'SnnsR'
setUnitDefaults(act, bias, st, subnet_no, layer_no, act_func, out_func)

Arguments

act
same as SNNS kernel function

bias
idem

st
idem

subnet_no
idem

layer_no
idem

act_func
idem

out_func
idem

[Package RSNNS version 0.4-7 Index]