SnnsRObject$somPredictCurrPatSetWinners {RSNNS}	R Documentation

Get most of the relevant results from a som

Description

SnnsR low-level function to get most of the relevant results from a SOM.

Usage

## S4 method for signature 'SnnsR'
somPredictCurrPatSetWinners(updateFuncParams=c(0.0, 0.0, 1.0),
saveWinnersPerPattern=TRUE, targets=NULL)

Arguments

updateFuncParams
parameters passed to the networks update function

saveWinnersPerPattern
should a list with the winners for every pattern be saved?

targets
optional target classes of the patterns

Value

a list with three elements:

nWinnersPerUnit
For each unit, the amount of patterns where this unit won is given. So, this is a 1d vector representing the normal version of the som.

winnersPerPattern
a vector where for each pattern the number of the winning unit is given. This is an intermediary result that normally won't be saved.

labeledUnits
a matrix which – if the targets parameter is given – contains for each unit (rows) and each class present in the targets (columns), the amount of patterns of the class where the unit has won. From the labeledUnits, the labeledMap can be computed, e.g. by voting of the class labels for the final label of the unit.

See Also

som

[Package RSNNS version 0.4-7 Index]