SnnsRObject$somPredictCurrPatSetWinnersSpanTree {RSNNS}	R Documentation

Get the spanning tree of the SOM

Description

SnnsR low-level function to get the spanning tree of the SOM, This function calls directly the corresponding SNNS kernel function (the only one available for SOM). Advantage are faster computation, disadvantage is somewhat limited information in the output.

Usage

## S4 method for signature 'SnnsR'
somPredictCurrPatSetWinnersSpanTree()

Value

the spanning tree, which is the som, showing for each unit a number identifying the last pattern for which this unit won. (We note that, also if there are more than one patterns, only the last one is saved)

See Also

som

[Package RSNNS version 0.4-7 Index]