SnnsRObject$whereAreResults {RSNNS}	R Documentation

Get a list of output units of a net

Description

SnnsR low-level function to get a list of output units of a net.

Usage

## S4 method for signature 'SnnsR'
whereAreResults(outputMethod="output")

Arguments

outputMethod
a string defining the output method of the net. Possible values are: "art1", "art2", "artmap", "assoz", "som", "output".

Details

Depending on the network architecture, output is present in hidden units, in output units, etc. In some network types, the output units have a certain name prefix in SNNS. This function finds the output units according to certain network types. The type is specified by outputMethod. If the given outputMethod is unknown, the function defaults to "output".

Value

a list of numbers identifying the units

[Package RSNNS version 0.4-7 Index]