SnnsRObjectFactory {RSNNS}	R Documentation

SnnsR object factory

Description

Object factory to create a new object of type SnnsR-class.

Usage

SnnsRObjectFactory()

Details

This function creates a new object of class SnnsR-class, initializes its only slot variables with a new environment, generates a new C++ object of class SnnsCLib, and an empty object serialization.

See Also

$, SnnsR-class

Examples

mySnnsObject <- SnnsRObjectFactory()
mySnnsObject$setLearnFunc('Quickprop')
mySnnsObject$setUpdateFunc('Topological_Order')

[Package RSNNS version 0.4-7 Index]