denormalizeData {RSNNS}	R Documentation

Revert data normalization

Description

Column-wise normalization of the input matrix is reverted, using the given parameters.

Usage

denormalizeData(x, normParams)

Arguments

x
input data

normParams
the parameters generated by an earlier call to normalizeData that will be used for reverting normalization

Details

The input matrix is column-wise denormalized using the parameters given by normParams. E.g., if normParams contains mean and sd for every column, the values are multiplied by sd and the mean is added

Value

column-wise denormalized input

See Also

normalizeData, getNormParameters

Examples

data(iris)
values <- normalizeData(iris[,1:4])
denormalizeData(values, getNormParameters(values))

[Package RSNNS version 0.4-7 Index]