encodeClassLabels {RSNNS}	R Documentation

Encode a matrix of (decoded) class labels

Description

Applies analyzeClassification row-wise to a matrix.

Usage

encodeClassLabels(x, method = "WTA", l = 0, h = 0)

Arguments

x
inputs

method
see analyzeClassification

l
idem

h
idem

Value

a numeric vector, each number represents a different class. A zero means that no class was assigned to the pattern.

See Also

analyzeClassification

Examples

data(iris)
labels <- decodeClassLabels(iris[,5])
encodeClassLabels(labels)

[Package RSNNS version 0.4-7 Index]