exportToSnnsNetFile {RSNNS}	R Documentation

Export the net to a file in the original SNNS file format

Description

Export the net that is present in the rsnns object in the original (.net) SNNS file format.

Usage

exportToSnnsNetFile(object, filename, netname = "RSNNS_untitled")

Arguments

object
the rsnns object

filename
path and filename to be written to

netname
name that is given to the network in the file

[Package RSNNS version 0.4-7 Index]