extractNetInfo {RSNNS}	R Documentation

Extract information from a network

Description

This function generates a list of data.frames containing the most important information that defines a network, in a format that is easy to use. To get the full definition in the original SNNS format, use summary.rsnns or exportToSnnsNetFile instead.

Usage

extractNetInfo(object)

Arguments

object
the rsnns object

Details

Internally, a call to SnnsRObject$extractNetInfo is done, and the results of this call are returned.

Value

a list containing information extracted from the network (see SnnsRObject$extractNetInfo).

See Also

SnnsRObject$extractNetInfo

[Package RSNNS version 0.4-7 Index]