getNormParameters {RSNNS}	R Documentation

Get normalization parameters of the input data

Description

Get the normalization parameters that are appended by normalizeData as attributes to the input data.

Usage

getNormParameters(x)

Arguments

x
input data

Details

This function is equivalent to calling attr(x, "normParams").

Value

the parameters generated by an earlier call to normalizeData

See Also

normalizeData, denormalizeData

[Package RSNNS version 0.4-7 Index]