inputColumns {RSNNS}	R Documentation

Get the columns that are inputs

Description

This function extracts all columns from a matrix whose column names begin with "in". The example data of this package follows this naming convention.

Usage

inputColumns(patterns)

Arguments

patterns
matrix or data.frame containing the patterns

[Package RSNNS version 0.4-7 Index]