matrixToActMapList {RSNNS}	R Documentation

Convert matrix of activations to activation map list

Description

Organize a matrix containing 1d vectors of network activations as 2d maps.

Usage

matrixToActMapList(m, nrow = 0, ncol = 0)

Arguments

m
the matrix containing one activation pattern in every row

nrow
number of rows the resulting matrices will have

ncol
number of columns the resulting matrices will have

Details

The input to this function is a matrix containing in each row an activation pattern/output of a neural network. This function uses vectorToActMap to reorganize the matrix to a list of matrices, whereby each row of the input matrix is converted to a matrix in the output list.

Value

a list containing the activation map matrices

See Also

vectorToActMap plotActMap

[Package RSNNS version 0.4-7 Index]