normalizeData {RSNNS}	R Documentation

Data normalization

Description

The input matrix is column-wise normalized.

Usage

normalizeData(x, type = "norm")

Arguments

x
input data

type

either

type string specifying the type of normalization. Implemented are "0_1", "center", and "norm"

or

attribute list of a former call to this method to apply e.g. normalization of the training data to the test data

Details

The parameter type specifies, how normalization takes place:

0_1

values are normalized to the [0,1]-interval. The minimum in the data is mapped to zero, the maximum to one.

center

the data is centered, i.e. the mean is substracted

norm

the data is normalized to mean zero, variance one

Value

column-wise normalized input. The normalization parameters that were used for the normalization are present as attributes of the output. They can be obtained with getNormParameters.

See Also

denormalizeData, getNormParameters

[Package RSNNS version 0.4-7 Index]