plotIterativeError {RSNNS}	R Documentation

Plot iterative errors of an rsnns object

Description

Plot the iterative training and test error of the net of this rsnns object.

Plot the iterative training and test error of the net of this rsnns object.

Usage

plotIterativeError(object, ...)

## S3 method for class 'rsnns'
plotIterativeError(object, ...)

Arguments

object
a rsnns object

...
parameters passed to plot

Details

Plots (if present) the class members IterativeFitError (as black line) and IterativeTestError (as red line).

[Package RSNNS version 0.4-7 Index]