plotRegressionError {RSNNS}	R Documentation

Plot a regression error plot

Description

The plot shows target values on the x-axis and fitted/predicted values on the y-axis. The optimal fit would yield a line through zero with gradient one. This optimal line is shown in black color. A linear fit to the actual data is shown in red color.

Usage

plotRegressionError(targets, fits, ...)

Arguments

targets
the target values

fits
the values predicted/fitted by the model

...
parameters passed to plot

[Package RSNNS version 0.4-7 Index]