readPatFile {RSNNS}	R Documentation

Load data from a pat file

Description

This function generates an SnnsR-class object, loads the given .pat file there as a pattern set and then extracts the patterns to a matrix, using SnnsRObject$extractPatterns.

Usage

readPatFile(filename)

Arguments

filename
the name of the .pat file

Value

a matrix containing the data loaded from the .pat file.

[Package RSNNS version 0.4-7 Index]