rsnnsObjectFactory {RSNNS}	R Documentation

Object factory for generating rsnns objects

Description

The object factory generates an rsnns object and initializes its member variables with the values given as parameters. Furthermore, it generates an object of SnnsR-class. Later, this information is to be used to train the network.

Usage

rsnnsObjectFactory(subclass, nInputs, maxit, initFunc, initFuncParams,
  learnFunc, learnFuncParams, updateFunc, updateFuncParams,
  shufflePatterns = TRUE, computeIterativeError = TRUE, pruneFunc = NULL,
  pruneFuncParams = NULL)

Arguments

subclass
the subclass of rsnns to generate (vector of strings)

nInputs
the number of inputs the network will have

maxit
maximum of iterations to learn

initFunc
the initialization function to use

initFuncParams
the parameters for the initialization function

learnFunc
the learning function to use

learnFuncParams
the parameters for the learning function

updateFunc
the update function to use

updateFuncParams
the parameters for the update function

shufflePatterns
should the patterns be shuffled?

computeIterativeError
should the error be computed in every iteration?

pruneFunc
the pruning function to use

pruneFuncParams
the parameters for the pruning function. Unlike the other functions, these have to be given in a named list. See the pruning demos for further explanation.

Details

The typical procedure implemented in rsnns subclasses is the following:

generate the rsnns object with this object factory

generate the network according to the architecture needed

train the network (with train)

In every rsnns object, the iterative error is the summed squared error (SSE) of all patterns. If the SSE is computed on the test set, then it is weighted to take care of the different amount of patterns in the sets.

Value

a partly initialized rsnns object

See Also

mlp, dlvq, rbf, rbfDDA, elman, jordan, som, art1, art2, artmap, assoz

[Package RSNNS version 0.4-7 Index]