savePatFile {RSNNS}	R Documentation

Save data to a pat file

Description

This function generates an SnnsR-class object, loads the given data there as a pattern set and then uses the functionality of SNNS to save the data as a .pat file.

Usage

savePatFile(inputs, targets, filename)

Arguments

inputs
a matrix with input values

targets
a matrix with target values

filename
the name of the .pat file

[Package RSNNS version 0.4-7 Index]