train {RSNNS}	R Documentation

Internal generic train function for rsnns objects

Description

The function calls SnnsRObject$train and saves the result in the current rsnns object. This function is used internally by the models (e.g. mlp) for training. Unless you are not about to implement a new model on the S3 layer you most probably don't want to use this function.

Internal generic train function for rsnns objects.

Usage

train(object, ...)

## S3 method for class 'rsnns'
train(object, inputsTrain, targetsTrain = NULL,
  inputsTest = NULL, targetsTest = NULL, serializeTrainedObject = TRUE,
  ...)

Arguments

object
the rsnns object

...
additional function parameters (currently not used)

inputsTrain
training input

targetsTrain
training targets

inputsTest
test input

targetsTest
test targets

serializeTrainedObject
parameter passed to SnnsRObject$train

Value

an rsnns object, to which the results of training have been added.

[Package RSNNS version 0.4-7 Index]