vectorToActMap {RSNNS}	R Documentation

Convert a vector to an activation map

Description

Organize network activation as 2d map.

Usage

vectorToActMap(v, nrow = 0, ncol = 0)

Arguments

v
the vector containing the activation pattern

nrow
number of rows the resulting matrices will have

ncol
number of columns the resulting matrices will have

Details

The input to this function is a vector containing in each row an activation pattern/output of a neural network. This function reorganizes the vector to a matrix. Normally, only the number of rows nrow will be used.

Value

a matrix containing the 2d reorganized input

See Also

matrixToActMapList plotActMap

[Package RSNNS version 0.4-7 Index]