confusionMatrix {RSNNS}	R Documentation

Computes a confusion matrix

Description

The confusion matrix shows how many times a pattern with the real class x was classified as class y. A perfect method should result in a diagonal matrix. All values not on the diagonal are errors of the method.

Usage

confusionMatrix(targets, predictions)

Arguments

targets
the known, correct target values

predictions
the corresponding predictions of a method for the targets

Details

If the class labels are not already encoded, they are encoded using encodeClassLabels (with default values).

Value

the confusion matrix

[Package RSNNS version 0.4-7 Index]