summary.rsnns {RSNNS}	R Documentation

Generic summary function for rsnns objects

Description

Prints out a summary of the network. The printed information can be either all information of the network in the original SNNS file format, or the information given by extractNetInfo. This behaviour is controlled with the parameter origSnnsFormat.

Usage

## S3 method for class 'rsnns'
summary(object, origSnnsFormat = TRUE, ...)

Arguments

object
the rsnns object

origSnnsFormat
show data in SNNS's original format in which networks are saved, or show output of extractNetInfo

...
additional function parameters (currently not used)

Value

Either the contents of the .net file that SNNS would generate from the object, as a string. Or the output of extractNetInfo.

See Also

extractNetInfo

[Package RSNNS version 0.4-7 Index]